Create search indexer to handle user reindexing of different implementation
