+++ pkg/lib/search/reindex/deps.go
@@ -1,0 +1,10 @@
+package reindex
+
+import (
+	"github.com/google/wire"
+)
+
+var DependencySet = wire.NewSet(
+	NewReindexerLogger,
+	wire.Struct(new(Reindexer), "*"),
+)
