+++ pkg/redisqueue/task_user_reindex.go
@@ -5,18 +5,18 @@ import (
 	"encoding/json"
 
 	"github.com/authgear/authgear-server/pkg/lib/deps"
-	"github.com/authgear/authgear-server/pkg/lib/elasticsearch"
 	"github.com/authgear/authgear-server/pkg/lib/infra/redisqueue"
+	"github.com/authgear/authgear-server/pkg/lib/search/reindex"
 )
 
 func UserReindex(ctx context.Context, appProvider *deps.AppProvider, task *redisqueue.Task) (output json.RawMessage, err error) {
-	esService := newElasticsearchService(ctx, appProvider)
-	var request elasticsearch.ReindexRequest
+	reindexer := newSearchReindexer(ctx, appProvider)
+	var request reindex.ReindexRequest
 	err = json.Unmarshal(task.Input, &request)
 	if err != nil {
 		return
 	}
-	result := esService.ExecReindexUser(ctx, request)
+	result := reindexer.ExecReindexUser(ctx, request)
 	resultBytes, err := json.Marshal(result)
 	if err != nil {
 		return
